Value map combobox filter
