+++ src/core/valuemapmodel.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************
-                            valuemapmodel.cpp
+                            valuemapmodelbase.cpp
 
                               -------------------
               begin                : March 2019
@@ -16,114 +16,36 @@
  *                                                                         *
  ***************************************************************************/
 
-#include "qgsvaluemapfieldformatter.h"
 #include "valuemapmodel.h"
-
-#include <QDebug>
+#include "valuemapmodelbase.h"
 
 ValueMapModel::ValueMapModel( QObject *parent )
-  : QAbstractListModel( parent )
+  : QSortFilterProxyModel( parent )
+  , mSourceModel( new ValueMapModelBase( this ) )
 {
+  setSourceModel( mSourceModel );
+
+  setFilterRole( ValueRole );
+
+  connect( mSourceModel, &ValueMapModelBase::mapChanged, this, &ValueMapModel::mapChanged );
 }
 
 QVariant ValueMapModel::map() const
 {
-  return mConfiguredMap;
+  return mSourceModel->map();
 }
 
 void ValueMapModel::setMap( const QVariant &map )
 {
-  mMap.clear();
-  // QGIS 3
-  const QVariantList list = map.toList();
-  if ( !list.empty() )
-  {
-    beginInsertRows( QModelIndex(), 0, list.size() );
-
-    for ( const QVariant &item : list )
-    {
-      const QVariantMap mapItem = item.toMap();
-
-      const QString key = mapItem.firstKey();
-      const QVariant value = mapItem.value( key );
-
-      mMap.append( qMakePair( value, key ) );
-    }
-    endInsertRows();
-  }
-  else // QGIS 2 compat
-  {
-    const QVariantMap valueMap = map.toMap();
-    if ( !valueMap.empty() )
-    {
-      beginInsertRows( QModelIndex(), 0, valueMap.size() );
-
-      QMapIterator<QString, QVariant> i( valueMap );
-      while ( i.hasNext() )
-      {
-        i.next();
-        const QString key = i.key();
-        const QVariant value = i.value();
-
-        mMap.append( qMakePair( value, key ) );
-      }
-      endInsertRows();
-    }
-  }
-
-  mConfiguredMap = map;
-  emit mapChanged();
-}
-
-int ValueMapModel::rowCount( const QModelIndex &parent ) const
-{
-  Q_UNUSED( parent )
-  return mMap.size();
-}
-
-QVariant ValueMapModel::data( const QModelIndex &index, int role ) const
-{
-  if ( role == KeyRole )
-    return mMap.at( index.row() ).first;
-  else
-    return mMap.at( index.row() ).second;
-}
-
-QHash<int, QByteArray> ValueMapModel::roleNames() const
-{
-  QHash<int, QByteArray> roles = QAbstractItemModel::roleNames();
-
-  roles[KeyRole] = "key";
-  roles[ValueRole] = "value";
-
-  return roles;
+  mSourceModel->setMap( map );
 }
 
 int ValueMapModel::keyToIndex( const QVariant &key ) const
 {
-  int i = 0;
-  for ( const auto &item : mMap )
-  {
-    if ( item.first.toString() == key.toString() )
-    {
-      return i;
-    }
-    ++i;
-  }
-  return -1;
+  return mSourceModel->keyToIndex( key );
 }
 
 QVariant ValueMapModel::keyForValue( const QString &value ) const
 {
-  QVariant result;
-
-  auto match = std::find_if( mMap.begin(), mMap.end(), [&value]( const QPair<QVariant, QString> &x ) { return x.second == value; } );
-
-  if ( match != mMap.end() )
-    result = match->first;
-
-  if ( result == QgsValueMapFieldFormatter::NULL_VALUE )
-    result = QVariant();
-
-  return result;
+  return mSourceModel->keyForValue( value );
 }
